fix(OUT-839) : applied requested changes
