+++ package.json
@@ -57,7 +57,7 @@
     "react-dnd-touch-backend": "^16.0.1",
     "react-dom": "^18",
     "react-redux": "^9.1.0",
-    "tapwrite": "^1.1.7",
+    "tapwrite": "^1.1.8",
     "zod": "^3.23.8"
   },
   "devDependencies": {
