+++ src/app/api/scrap-images/route.ts
@@ -1,0 +1,4 @@
+import { withErrorHandler } from '@api/core/utils/withErrorHandler'
+import { PostScrapImage } from '@api/scrap-images/scrap-images.controller'
+
+export const POST = withErrorHandler(PostScrapImage)
